+++ HS2_Sideloader/Properties/AssemblyInfo.cs
@@ -1,0 +1,4 @@
+﻿using System.Reflection;
+
+[assembly: AssemblyProduct("Sideloader")]
+[assembly: AssemblyTitle("BepisPlugins.Sideloader")]
