+++ HS2_Sideloader/packages.config
@@ -1,0 +1,16 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="IllusionLibs.BepInEx" version="5.0.1.1" targetFramework="net46" developmentDependency="true" />
+  <package id="IllusionLibs.BepInEx.Harmony" version="2.0.0-beta2" targetFramework="net46" developmentDependency="true" />
+  <package id="IllusionLibs.HoneySelect2.Assembly-CSharp" version="2020.5.29" targetFramework="net46" developmentDependency="true" />
+  <package id="IllusionLibs.HoneySelect2.Assembly-CSharp-firstpass" version="2020.5.29" targetFramework="net46" developmentDependency="true" />
+  <package id="IllusionLibs.HoneySelect2.IL" version="2020.5.29" targetFramework="net46" developmentDependency="true" />
+  <package id="IllusionLibs.HoneySelect2.UnityEngine.AssetBundleModule" version="2018.4.11" targetFramework="net46" developmentDependency="true" />
+  <package id="IllusionLibs.HoneySelect2.UnityEngine.AudioModule" version="2018.4.11" targetFramework="net46" developmentDependency="true" />
+  <package id="IllusionLibs.HoneySelect2.UnityEngine.CoreModule" version="2018.4.11" targetFramework="net46" developmentDependency="true" />
+  <package id="IllusionLibs.HoneySelect2.UnityEngine.ImageConversionModule" version="2018.4.11" targetFramework="net46" developmentDependency="true" />
+  <package id="IllusionLibs.HoneySelect2.UnityEngine.UnityWebRequestWWWModule" version="2018.4.11" targetFramework="net46" developmentDependency="true" />
+  <package id="SharpZipLib" version="1.2.0" targetFramework="net46" />
+  <package id="XUnity.Common" version="1.0.0" targetFramework="net46" />
+  <package id="XUnity.ResourceRedirector" version="1.1.0" targetFramework="net46" />
+</packages>
