+++ src/Core_Sideloader/Core.Sideloader.ListLoader.Hooks.cs
@@ -1,6 +1,6 @@
 ﻿using BepInEx.Harmony;
 using HarmonyLib;
-#if AI
+#if AI || HS2
 using AIChara;
 #endif
