+++ src/Core_Sideloader/Core_Sideloader.projitems
@@ -10,6 +10,9 @@
   </PropertyGroup>
   <ItemGroup>
     <Compile Include="$(MSBuildThisFileDirectory)Core.Sideloader.ListLoader.Hooks.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)UniversalAutoResolver\Core.UAR.StructReference.AI.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)UniversalAutoResolver\Core.UAR.FaceSkinInfo.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)UniversalAutoResolver\Core.UAR.HeadPresetInfo.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)UniversalAutoResolver\Core.UAR.MigrationInfo.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Core.Sideloader.ReflectionHelper.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Core.Sideloader.ListLoader.cs" />
