+++ src/Core_Sideloader/UniversalAutoResolver/Core.UAR.MigrationInfo.cs
@@ -1,4 +1,4 @@
-﻿#if AI
+﻿#if AI || HS2
 using AIChara;
 #endif
