+++ src/Core_Sideloader/UniversalAutoResolver/Core.UAR.ResolveInfo.cs
@@ -1,6 +1,6 @@
 ﻿using MessagePack;
 using System;
-#if AI
+#if AI || HS2
 using AIChara;
 #endif
