+++ src/Core_Sideloader/UniversalAutoResolver/Core.UAR.cs
@@ -5,7 +5,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading;
 using Logging = BepInEx.Logging;
-#if AI
+#if AI || HS2
 using AIChara;
 #endif
 
@@ -29,6 +29,11 @@ namespace Sideloader.AutoResolver
         private static ILookup<int, ResolveInfo> _resolveInfoLookupLocalSlot;
         private static ILookup<string, MigrationInfo> _migrationInfoLookupGUID;
         private static ILookup<int, MigrationInfo> _migrationInfoLookupSlot;
+#if AI || HS2
+        private static ILookup<int, HeadPresetInfo> _headPresetInfoLookupSlot;
+        private static ILookup<int, FaceSkinInfo> _faceSkinInfoLookupSlot;
+        private static ILookup<int, FaceSkinInfo> _faceSkinInfoLookupLocalSlot;
+#endif
 
         /// <summary>
         /// The starting point for UAR IDs
@@ -106,6 +111,17 @@ namespace Sideloader.AutoResolver
         /// <returns>A list of MigrationInfo</returns>
         public static List<MigrationInfo> GetMigrationInfo(int idOld) => _migrationInfoLookupSlot?[idOld].ToList();
 
+#if AI || HS2
+        internal static HeadPresetInfo TryGetHeadPresetInfo(int slot, string guid, string preset) =>
+            _headPresetInfoLookupSlot?[slot].FirstOrDefault(x => x.HeadGUID == guid && x.Preset == preset);
+        internal static void SetHeadPresetInfos(ICollection<HeadPresetInfo> results) => _headPresetInfoLookupSlot = results.ToLookup(info => info.HeadID);
+
+        internal static FaceSkinInfo TryGetFaceSkinInfo(int slot, string guid) =>
+            _faceSkinInfoLookupSlot?[slot].FirstOrDefault(x => x.SkinSlot == slot && x.SkinGUID == guid);
+        internal static FaceSkinInfo TryGetFaceSkinInfo(int localSlot) =>
+            _faceSkinInfoLookupLocalSlot?[localSlot].FirstOrDefault();
+#endif
+
         internal static void SetResolveInfos(ICollection<ResolveInfo> results)
         {
             _resolveInfoLookupSlot = results.ToLookup(info => info.Slot);
@@ -116,6 +132,21 @@ namespace Sideloader.AutoResolver
             _migrationInfoLookupGUID = results.ToLookup(info => info.GUIDOld);
             _migrationInfoLookupSlot = results.ToLookup(info => info.IDOld);
         }
+
+#if AI || HS2
+        internal static void SetFaceSkinInfos(ICollection<FaceSkinInfo> results)
+        {
+            foreach (var info in results)
+            {
+                var resolveInfo = TryGetResolutionInfo(info.SkinSlot, ChaListDefine.CategoryNo.ft_skin_f, info.SkinGUID);
+                if (resolveInfo != null)
+                    info.SkinLocalSlot = resolveInfo.LocalSlot;
+            }
+
+            _faceSkinInfoLookupSlot = results.ToLookup(info => info.SkinSlot);
+            _faceSkinInfoLookupLocalSlot = results.ToLookup(info => info.SkinLocalSlot);
+        }
+#endif
 
         /// <summary>
         /// Change the ID of items saved to a card to their resolved IDs
@@ -169,13 +200,13 @@ namespace Sideloader.AutoResolver
                 {
 #if KK || EC
                     if (Lists.InternalDataList[kv.Key.Category].ContainsKey(kv.Value.GetMethod(structure)))
-#elif AI
+#elif AI || HS2
                     if (Lists.InternalDataList[(int)kv.Key.Category].ContainsKey(kv.Value.GetMethod(structure)))
 #endif
                     {
 #if KK || EC
                         string mainAB = Lists.InternalDataList[kv.Key.Category][kv.Value.GetMethod(structure)].dictInfo[(int)ChaListDefine.KeyType.MainAB];
-#elif AI
+#elif AI || HS2
                         string mainAB = Lists.InternalDataList[(int)kv.Key.Category][kv.Value.GetMethod(structure)].dictInfo[(int)ChaListDefine.KeyType.MainAB];
 #endif
                         mainAB = mainAB.Replace("chara/", "").Replace(".unity3d", "").Replace(kv.Key.Category.ToString() + "_", "").Replace("/", "");
@@ -184,7 +215,7 @@ namespace Sideloader.AutoResolver
                         {
                             //ID found but it conflicts with a vanilla item. Change the ID to avoid conflicts.
                             ShowGUIDError(extResolve.GUID);
-                            if (kv.Key.Category.ToString().Contains("ao_") && Sideloader.KeepMissingAccessories.Value && Manager.Scene.Instance.NowSceneNames.Any(sceneName => sceneName == "CustomScene"))
+                            if (kv.Key.Category.ToString().Contains("ao_") && Sideloader.KeepMissingAccessories.Value && GetNowSceneNames().Any(sceneName => sceneName == "CustomScene"))
                                 kv.Value.SetMethod(structure, 1);
                             else
                                 kv.Value.SetMethod(structure, BaseSlotID - 1);
@@ -199,7 +230,7 @@ namespace Sideloader.AutoResolver
                     {
                         //ID not found. Change the ID to avoid potential future conflicts.
                         ShowGUIDError(extResolve.GUID);
-                        if (kv.Key.Category.ToString().Contains("ao_") && Sideloader.KeepMissingAccessories.Value && Manager.Scene.Instance.NowSceneNames.Any(sceneName => sceneName == "CustomScene"))
+                        if (kv.Key.Category.ToString().Contains("ao_") && Sideloader.KeepMissingAccessories.Value && GetNowSceneNames().Any(sceneName => sceneName == "CustomScene"))
                             kv.Value.SetMethod(structure, 1);
                         else
                             kv.Value.SetMethod(structure, BaseSlotID - 1);
@@ -213,7 +244,7 @@ namespace Sideloader.AutoResolver
             //Only attempt compatibility resolve if the ID does not belong to a vanilla item or hard mod
 #if KK || EC
             if (!Lists.InternalDataList[kv.Key.Category].ContainsKey(kv.Value.GetMethod(structure)))
-#elif AI
+#elif AI || HS2
             if (!Lists.InternalDataList[(int)kv.Key.Category].ContainsKey(kv.Value.GetMethod(structure)))
 #endif
             {
@@ -248,7 +279,7 @@ namespace Sideloader.AutoResolver
                 //No match was found
                 if (Sideloader.DebugLogging.Value)
                     Sideloader.Logger.LogDebug($"Compatibility resolving failed, no match found for ID {kv.Value.GetMethod(structure)} Category {kv.Key.Category}");
-                if (kv.Key.Category.ToString().Contains("ao_") && Sideloader.KeepMissingAccessories.Value && Manager.Scene.Instance.NowSceneNames.Any(sceneName => sceneName == "CustomScene"))
+                if (kv.Key.Category.ToString().Contains("ao_") && Sideloader.KeepMissingAccessories.Value && GetNowSceneNames().Any(sceneName => sceneName == "CustomScene"))
                     kv.Value.SetMethod(structure, 1);
             }
         }
@@ -258,7 +289,7 @@ namespace Sideloader.AutoResolver
             action(StructReference.ChaFileFaceProperties, file.custom.face, extInfo, "");
             action(StructReference.ChaFileBodyProperties, file.custom.body, extInfo, "");
             action(StructReference.ChaFileHairProperties, file.custom.hair, extInfo, "");
-#if AI
+#if AI || HS2
             action(StructReference.ChaFileMakeupProperties, file.custom.face.makeup, extInfo, "");
 #else
             action(StructReference.ChaFileMakeupProperties, file.custom.face.baseMakeup, extInfo, "");
@@ -288,6 +319,43 @@ namespace Sideloader.AutoResolver
                 action(StructReference.ChaFileAccessoryPartsInfoProperties, coordinate.accessory.parts[acc], extInfo, accPrefix);
             }
         }
+
+#if AI || HS2
+        internal static void ResolveFaceSkins()
+        {
+            foreach (var data in Lists.ExternalDataList.Where(x => x.categoryNo == (int)ChaListDefine.CategoryNo.ft_skin_f))
+            {
+                int IDIndex = data.lstKey.IndexOf("ID");
+                int headIDIndex = data.lstKey.IndexOf("HeadID");
+
+                foreach (var x in data.dictList)
+                {
+                    int id = int.Parse(x.Value[IDIndex]);
+                    int headID = int.Parse(x.Value[headIDIndex]);
+
+                    var faceSkinInfo = TryGetFaceSkinInfo(id);
+                    if (faceSkinInfo == null) continue;
+
+                    var resolveInfo = TryGetResolutionInfo(faceSkinInfo.HeadSlot, ChaListDefine.CategoryNo.fo_head, faceSkinInfo.HeadGUID);
+                    if (resolveInfo == null)
+                        ShowGUIDError(faceSkinInfo.HeadGUID);
+                    else
+                    {
+                        if (headID != faceSkinInfo.HeadSlot)
+                        {
+                            Sideloader.Logger.LogError($"Error resolving face skins, head ID in manifest does do not match ID in list for GUID:{faceSkinInfo.SkinGUID}, skin ID:{faceSkinInfo.SkinSlot}");
+                            continue;
+                        }
+
+                        if (Sideloader.DebugLoggingResolveInfo.Value)
+                            Sideloader.Logger.LogDebug($"Resolving face skin ({faceSkinInfo.SkinGUID}) head ID ({faceSkinInfo.HeadGUID}) from slot {faceSkinInfo.HeadSlot} to slot {resolveInfo.LocalSlot}");
+
+                        x.Value[headIDIndex] = resolveInfo.LocalSlot.ToString();
+                    }
+                }
+            }
+        }
+#endif
 
         internal static void MigrateData(ResolveInfo extResolve, ChaListDefine.CategoryNo categoryNo)
         {
@@ -394,6 +462,20 @@ namespace Sideloader.AutoResolver
             results.AddRange(manifest.MigrationList);
         }
 
+#if AI || HS2
+        internal static void GenerateHeadPresetInfo(Manifest manifest, List<HeadPresetInfo> results)
+        {
+            manifest.LoadHeadPresetInfo();
+            results.AddRange(manifest.HeadPresetList);
+        }
+
+        internal static void GenerateFaceSkinInfo(Manifest manifest, List<FaceSkinInfo> results)
+        {
+            manifest.LoadFaceSkinInfo();
+            results.AddRange(manifest.FaceSkinList);
+        }
+#endif
+
         internal static void ShowGUIDError(string guid)
         {
             Logging.LogLevel loglevel = Sideloader.MissingModWarning.Value ? Logging.LogLevel.Warning | Logging.LogLevel.Message : Logging.LogLevel.Warning;
@@ -401,7 +483,7 @@ namespace Sideloader.AutoResolver
             if (LoadedResolutionInfo.Any(x => x.GUID == guid))
                 //we have the GUID loaded, so the user has an outdated mod
                 Sideloader.Logger.Log(loglevel, $"[UAR] WARNING! Outdated mod detected! [{guid}]");
-#if KK || AI
+#if KK || AI || HS2
             else if (LoadedStudioResolutionInfo.Any(x => x.GUID == guid))
                 //we have the GUID loaded, so the user has an outdated mod
                 Sideloader.Logger.Log(loglevel, $"[UAR] WARNING! Outdated mod detected! [{guid}]");
@@ -409,6 +491,15 @@ namespace Sideloader.AutoResolver
             else
                 //did not find a match, we don't have the mod
                 Sideloader.Logger.Log(loglevel, $"[UAR] WARNING! Missing mod detected! [{guid}]");
+        }
+
+        private static List<string> GetNowSceneNames()
+        {
+#if HS2
+            return Manager.Scene.NowSceneNames;
+#else
+            return Manager.Scene.Instance.NowSceneNames;
+#endif
         }
     }
 }
